requirements.txt
PySide6
python-pptx
openai
python-dotenv
requests
PyInstaller
